requirements.txt
streamlit==1.43.2
python-box==7.3.2
transformers==4.49.0
ensure==1.0.4
PyYAML==6.0.2
datasets==3.4.1
torch==2.2.2
bitsandbytes==0.42.0
peft==0.15.0
accelerate==1.5.2
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv=1.1.0
pinecone==6.0.2
openai==1.68.2
langchain==0.3.21
langchain-huggingface==0.1.2
sentence-transformers==3.4.1
-e .